Refactor #24 : PR 내용 수정
