+++ app/(my-page)/my-page/product-list/actions.ts
@@ -1,16 +1,17 @@
 "use server";
 import { fireStore } from "@/config/firebase/firebase";
 import { collection, getDocs } from "firebase/firestore";
-import { Product } from "./page";
+import { IProduct } from "@/model/product";
+import { COLLECTION_NAME_PRODUCT } from "@/constants/variables";
 
-export async function getProduct(): Promise<Product[]> {
-	const productsCol = collection(fireStore, "Product");
+export async function getProduct(): Promise<IProduct[]> {
+	const productsCol = collection(fireStore, COLLECTION_NAME_PRODUCT);
 
 	try {
 		const productSnapshot = await getDocs(productsCol);
 		const productList = productSnapshot.docs.map((doc) => ({
 			id: doc.id,
-			...(doc.data() as Omit<Product, "id">),
+			...(doc.data() as Omit<IProduct, "id">),
 		}));
 		// console.log("Product List : ", productList);
 		return productList;
